requirements.txt
xarray==2023.4.2
fastapi==0.95.1
matplotlib==3.7.1
pydantic==1.10.7
numpy==1.24.3
pillow==9.5.0
uvicorn==0.20.0
netcdf4==1.6.3
netcdf==66.0.2